Backport of Intuos4 LED and OLED control
The "Intous4 LED and OLED control" patch has been accepted into
the linux-input tree, and is scheduled for inclusion into linux-3.2.
This is a backport of said patch into the wacom-input tree in order
to support this feature also on older kernels. The ABI description
is not contained in this patch, because the authoritative source
of the ABI should stay in the linux kernel.

Additionally, the usb_set_report retransmission handling is not
included in this backport patch, because
* it has been found that retransmission is not needed, and
* adding it would make the patch less concise

This commit enables control of the LEDs and OLED displays found on the
Wacom Intuos4 M, L, and XL. For this purpose, a new "wacom_led" attribute
group is added to the sysfs entry of the USB device.

This "wacom_led" group only shows up when the correct device (M, L, or XL)
is detected.

Signed-off-by: Eduard Hasenleithner <eduard@hasenleithner.at>
Signed-off-by: Ping Cheng <pinglinux@gmail.com>