Input: wacom - Mark Intuos5 pad as in-prox when touching buttons
If the ExpressKeys on the Intuos5 are touched, they currently send
an out-of-prox event even if they're already out of prox. This can
cause some confusion in the X driver, which doesn't expect out-of-
prox notification for an already out-of-prox tool. To fix this, we
send an in-prox notification when the buttons are touched, despite
not sending any data to go along with that notification.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>