Fix spurious SOURCE_STYLUS attribute for MFT under Android
Under Android, the MFT sensor of the Cintiq Companion Hybrid is
being assigned a SOURCE_STYLUS attribute, in addition to its
expected SOURCE_TOUCHSCREEN. This is caused by the
MultiTouchMotionAccumulator (InputReader.cpp) assuming that a
stylus may be used with any device advertising the
ABS_MT_TOOL_TYPE axis.

This patch removes our advertising of that particular axis for
the MFT sensor.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Acked-by: Chris Bagwell <chris@cnpbagwell.com>