Fix Intuos5 and Intuos Pro touch interfaces on 2.6.30+
The touch interfaces on for these two devices have not been set
up properly and will likely be mis-detected by xf86-input-wacom
and result in a non-functional touch device. This patch ensures
the driver declares the appropriate axes for the pen and touch
interfaces, and routes touches from both through the appropriate
IRQ method.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Acked-by: Chris Bagwell <chris@cnpbagwell.com>