Declare ABS_X and ABS_Y for all device types on 2.6.38
Devices capable of >2 touches were not declaring support for the ABS_X
and ABS_Y axes despite using pointer emulation. The 3.7+ kernel already
takes care of this for us (so there's no need to forward-port these
changes), but we need to be sure that they are declared under 2.6.38+
lest xf86-input-wacom ignore the device.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Acked-by: Chris Bagwell <chris@cnpbagwell.com>