Input: wacom - override 'pressure_max' with value from HID_USAGE_PRESSURE
The 0xEC sensor is used in multiple tablet PCs and curiously has versions
that report 256 levels of pressure (Samsung Slate 7) as well as versions
that report 1024 levels (Lenovo Thinkpad Yoga).  To allow both versions to
work properly, we allow the value of HID_USAGE_PRESSURE reported to
override pressure_max.

 ** HISTORIC NOTE **

In the 2.6.30 tree, we had some basic pressure support. However,
according to the comments, there wasn't actually a reported
maximum. We currently expect a logical maximum following this,
but that apparently doesn't always hold. It was forced to 255
here.

Also, apparently fingers used to report the amount of sensed
capacitance (a reasonable indicator of pressure) and included
info in the HID descriptor. (HID_USAGE_DESKTOP << 16 | HID_USAGE_UNDEFINED)
stored a 16-bit value at report[i+3] just like our current
pressure. This test was removed from the mainline a few years
ago because the the reported capacitance values were generally
too bogus to be useful.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>