Use backwards-compatible battery name only for first tablet
Attaching multiple wireless tablets to a system will cause a kernel
crash when both power supply nodes try to use the same name. The 3.17
kernel adds a unique suffix to the end of every device, but we can't
do that here if we want to be backwards compatible. Instead, we use the
suffix-less name only if it has not already been registered and use the
new behavior otherwise.

Fixes SF bug #248

Signed-off-by: Jason Gerecke <killertofu@gmail.com>