configure.ac: Do not search for headers when finding build directory
Some distros (I'm looking at you, SUSE) keep the kernel headers in a
directory separate from the build directory. If we end up trying to use
such a directory, the build will fail since the other necessary build
files don't exist.

The kernel makefile requires that a config file be present before it
will bother compiling. Check to make sure that one of them exists, and
assume that if the distro keeps the headers elsewhere that the kernel
will still be able to find them (which is the case on SUSE).

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>