configure.ac: Allow user to specify kernel version to compile for
This patch introduces a "--with-kernel-version" switch that allows the
user to specify the specific kernel version that they want to compile
the input-wacom modules for. This is especially useful to test that the
code compiles correctly for different kernels without the associated
requirement that those kernels be running at the moment.

To do this, we move the kernel version checks (and calculation of
WCM_KERNEL_VER) above the code to search for the kernel directory so
that it has an idea of where exactly to search. The version number
checks that used to set MODUTS are kept as a sanity check.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>