Fix power_supply issues under 4.1
The power supply interface changes in Linux 4.1, which requires to
changes to all drivers which make use of it. The upstream commit which
did this (297d716) was not on Jiri's branch, which is why we didn't
notice it until just now. In addition to the changes made by that patch,
Linus had to resolve several merge issues (8de29a3) caused by changes to
our driver's use of the power supply.

This patch rolls up both sets of changes into one and adds the necessary
preprocessor guards to keep things working with earlier kernels.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>