backport: HID: wacom: remove hardcoded WACOM_QUIRK_MULTI_INPUT
The quirk was added for devices that support both pen and touch.
It decides if a device supports multiple inputs by hardcoded
feature type. However, for some devices, we do not know if they
support both before accessing their HID descriptors.

This patch relies on dynamically assigned device_type to make
the decision. Also, we make it certain that wacom_wac->shared
is always created. That is, the driver will not be loaded if it
fails to create wacom_wac->shared.

Signed-off-by: Ping Cheng <pingc@wacom.com>