Run 'depmod -a' as part of the 'make install' process
Now that 'make install' does not replace the original wacom.ko or
wacom_w8001.ko file, users have to run 'depmod -a' to rebuild the
kernel's module dependencies before the updated modules will be
used. Because it is easy to forget this step and the exact command
may vary depending on the build target, include it as a step in the
install process.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>