Fixed Cintiq Companion 2 buttons.
The Companion 2 has two additional buttons compared to the Hybrid, and the positions of the data in the input data are offset by a couple of bytes. Device ID seems to be in the same location.

This change adds a CINTIQ_COMPANION_2 feature type to differentiate it from CINTIQ_HYBRID.

The button order for the directional pad was copied off of the Hybrid, using the four indices after the normal buttons, except for the center button, which is zero. This may not be ideal.