Disable distcheck rule to make release.sh happy
The new release.sh script runs 'make distcheck' by default to create the
tarballs that will be uploaded to the server. The 'distcheck' target
requires that the project support doing "VPATH builds" (a way of running
the build in a different source tree), but input-wacom is not capable of
them.

This patch causes the 'distcheck' target to simply run 'dist' instead.
This allows release.sh to complete successfully without having to pass
the '--dist' flag. When we get to the point of adding actual checks to
input-wacom, this hack will likely need to be removed in favor of
supporting VPATH builds.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>