Prioritize compiled input-wacom modules with a depmod.d entry
There have been repeated reports of the 'make install' step not working
as expected. After installing and rebooting, the built-in kernel module
is loaded in preference to the compiled input-wacom version. From what
little information I've been able to gather, the issue appears to be a
result of inconsistent priority between the built-in and 'extra'
directories.

To ensure the input-wacom module is always preferred when found, this
patch has the 'make install' step create a configuration file at
/etc/depmod.d/input-wacom.conf with an "override" directive. See the
'depmod.d' manpage for more information.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>