Provide uninstall target in Makefile
If a user would like to remove this externally-compiled module from their
system, it is now possible by running 'make uninstall'. Since the kernel
provides no target for this (unlike "modules_install") we simply delete
the modules from the "extra" subdirectory and update depmod so that the
kernel will load the distribution-supplied driver in the future.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>