Prevent hybrid mode crash of Cintiq Companion 2
When the Cintiq Companion 2 enters hybrid mode (i.e., is used as an
external display tablet), a series of kernel Oops (and an eventual
hang) occurs due to a null pointer dereference by 'hid_submit_ctrl'
in usbhid. The cause of the bug is under investigation, but changing
the touch interface to be of type CINTIQ_24HDT does seem to prevent
it from occuring.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>