Add support for Travis-CI integration
"Travis CI" is a continuous integration service which is made available
for free to open-source projects. When connected to a project at GitHub,
it is capable of automatically building projects and sending notifications
if a commit breaks the build. This commit adds the necessary .travis.yml
file and attempts to compile the driver against multiple different kernel
versions (one for each folder and conditional "LINUX_VERSION_CODE" codepath).

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>