Allow user to specify kernel directory or version independently
It should be possible to point the ./configure script at a kernel source
tree and have it work without needing to also specifiy exactly what
kernel version its looking at. This patch modifies configure.ac so that
the MODUTS is only set to the output of `uname -r` if neither a kernel
location nor version is given, and so the "sanity check" process which
compares the kernel version is able to set MODUTS if it hasn't already
been.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Tested-by: Ping Cheng <pingc@wacom.com>