Remove useless KERNEL_VERSION checks from 2.6.36 tree
The 2.6.36 tree includes several blocks of code that switch between
different statements depending on if you are compiling a kernel > 2.6.37
or not. Such kernels will not be compiled against the 2.6.36 codebase
though since they'll be directed to the "2.6.38" instead. Note that a
kernel with version e.g. 2.6.37.6 compares as <= 2.6.37 since only the
first three parts of the version are checked.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>