Fix deadlock on proximity in events with the Intuos Draw
Unfortunately, when we have an Intuos Draw connected using a USB
connection, wacom's IRQ handler will be called while the lock for the
usbhid driver is held by hid_ctrl. This means when we try to schedule a
new proximity event in wacom_intuos_schedule_prox_event() by calling
hid_hw_request(), we'll try to take the lock a second time which
deadlocks the system. This patch fixes that behavior by initializing a
worker when we have a INTUOSHT2 device connected, and using that worker
to schedule the proximity event instead.

NOTE: An alternative fix is due for inclusion in Linux 4.5 as commit e470127
("HID: usbhid: fix recursive deadlock") and fixes this issue by shrinking
the critical region to not interfere. These two fixes are compatible and
so aren't ifdef'd here.

Signed-off-by: Lyude <cpaul@redhat.com>
[jason.gerecke@wacom.com: Imported into input-wacom repository (not upstream)]
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>