Have git-version-gen produce better default output
Instead of producing "UNKNOWN" as its output when the real version
is not known, have git-version-gen produce the name of the directory
that the input-wacom sources are contained in. In some circumstances
(e.g. building from an archive generated by github) this can result
in useful information being added to the version string to help us
figure out exactly what is installed.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>