Run 'update-initramfs -u' during 'make install' if present
Recent Ubuntu releases (e.g. 15.10) include the Wacom driver in their
initramfs. If a tablet is connected at boot, this driver will be loaded
in preference to whatever we may have installed to /lib/modules. Since
we can't prevent this from occuring, we need to update the initramfs
to contain our module. Thankfully, Ubuntu provides a command to do just
that: update-initramfs. By running this after the module is installed
and depmod updated, the initramfs will be updated to contain our on-disk
driver and load it on boot.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>