backport: HID: wacom: Cleanup touch arbitration logic
stylus_in_proximity was introduced to support touch arbitration
before in range was supported. With in range event, the logic
changed. stylus_in_proximity should be set for both in prox and
in range events.

To finish a clean touch arbitration logic, we should send touch
up (if it was down) before posting any general pen events.

Signed-off-by: Ping Cheng <pingc@wacom.com>
Reviewed-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
[jason.gerecke@wacom.com: Imported into input-wacom repository (c1b03f5)]
Fixes: http://sourceforge.net/p/linuxwacom/bugs/301/
Fixes: linuxwacom/input-wacom#14
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
[aaron.skomra@wacom.com: backported from input-wacom repository (29b4a21)]
Signed-off-by: Aaron Skomra <aaron.skomra@wacom.com>