backport: MobileStudio Pro to 3.7, 2.6.38, 2.6.36, 2.6.30
Backports support for the MobileStudio Pro to all pre-3.17 kernels. Pen,
touch, and pad input should work on all kernels (though 2.6.30 is limited
to a maximum two-finger multitouch). Battery status reporting is limited
to 2.6.38 and newer. Accelerometer reporting is not possible with kernels
prior to 3.17 and so is not included here.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Acked-by: Ping Cheng <ping.cheng@wacom.com>