backport: HID: wacom: Rename wacom ID report ID macros
"INTUOSREAD" and "INTUOSWRITE" are poorly named. These are report IDs
for pen ID (proximity) packets. It should be noted that the latter is
only used on Intuos/Intuos2 for a second stylus when DualTrack is in use.

NOTE: This patch was backported earlier, but only as far back as 2.6.38.
To make backporting the next few patches easier, we complete this backport
all the way to 2.6.30.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
[jason.gerecke@wacom.com: Imported into input-wacom repository (0610999)]
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
[jason.gerecke@wacom.com: backported from input-wacom a56ed5b]
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
[jason.gerecke@wacom.com: backported from input-wacom (35dc103)]
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>