Properly clear 'touch_down' flag in 'wacom_mt_touch' for 2.6.38/3.7
The 'wacom_mt_touch' function is responsible for handling touch input from
some kinds of tablet PCs (those using report ID 13 or 3). This function
contains a logic error which causes the 'touch_down' flag to not be cleared
after the user stops touching the display. This could prevent the pen from
being usable on devices with "shared" pen and touch sensors.

This issue was corrected in the 3.17 tree (and by extension 4.5) as part of
import commit 04e1fb9 (upstream: e0d41fd) but was apparently overlooked
for the 2.6.38 and 3.7 trees during the backport 77249d4 (the 2.6.36 tree
received the fix, however).

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>