backport: HID: wacom: generic: support LEDs
Add support for the LEDs around the mode switch to the generic code path in
support of the second generation Intuos Pro.

Backport note: Kernel control of the LED mode switch is not present in
kernels before 4.5, and the WACOM_HID_WD_BUTTONCENTER usage is mapped,
but not handled specially by the driver here.

Signed-off-by: Aaron Skomra <aaron.skomra@wacom.com>
Reviewed-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
[aaron.skomra@wacom.com: Imported into input-wacom repository (10c55ca)]
Signed-off-by: Aaron Armstrong Skomra <aaron.skomra@wacom.com>
[aaron.skomra@wacom.com: Backported from input-wacom repository (fc3d6d5)]
Signed-off-by: Aaron Armstrong Skomra <aaron.skomra@wacom.com>