backport: Make wacom_msprot_irq more generic
The 'wacom_msprot_irq' function is currently used only for the MobileStudio
Pro, but could easily be made to handle the 2nd-generation Intuos Pro (or
other multitouch tablets) with a little bit of work. This patch modifies
the function to make it more easy to adapt for use by other devices.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>