backport: Fix handling of out-of-range packets in wacom_mspro_pen_irq
The pen reports sent by the MobileStudio Pro (and other tablets from
its generation) have a 'range' bit that indicates whether a tool is
sufficiently-close to the sensor to provide high-quality data. While
outside of "range" but in "prox", we need to be careful to not send data
to userspace since it may be low-quality or completely invalid. The state
of 'stylus_in_proximity' should remain 'true', however, to ensure that
palm rejection continues to work at these larger distances.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Aaron Armstrong Skomra <aaron.skomra@wacom.com>