2.6.30: Define sibling devices for MobileStudio Pro
The MobileStudio Pro, like other Cintiqs, has pen and touch data come
through entirely separately USB devices. The necessesity of specifying
an oVid and oPid was overlooked since it appeared that this version of
the driver did not support the properties. Now that proper multi-PID
device sharing is working (see prior patch), we define these properties.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>