Update Travis-CI configuration to use Ubuntu 14.04
It appears that the cdn.kernel.org server has recently stopped supporting
clients clients which try to connect with TLS 1.1[1]. This is a problem
for our Travis-CI tests which run in an Ubuntu 12.04 container (that
does not support later TLS versions) by default. There's no reason we
can't do our tests on a different container, and so explicitly tell
Travis to use their 14.04 "Trusty" container instead. This will let
our script once again download the kernel sources necessary to do the
build tests.

[1]: curl: (35) error:1407742E:SSL routines:SSL23_GET_SERVER_HELLO:tlsv1 alert protocol version

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>