Allow __wacom_is_usb_parent to handle unconfigured USB devices
The USB protocol allows devices to have multiple "configurations" that
the OS can switch between. A an "unconfigured" state is also possible,
especially when the device is first connected. If '__wacom_is_usb_parent'
comes across an unconfigured USB device, the 'config' variable will
be NULL. We need to add a check for this state before attempting a
dereference.

Fixes: 38c795d ("backport: HID: wacom: Improve generic name generation")
Fixes: bb55947 ("HID: wacom: Improve generic name generation")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wavom.com>