Synchronize trivial changes into to 3.17
A number of minor changes have either been unintentionally made to the
3.17 tree or occurred as a result of not backporting all upstream changes.
These differences do not affect driver operation, but the clutter they
produce in the output of `diff` does serve to hide a number of "real"
semantic differences. They also make it more difficult to backport
upstream patches, as the patch command will fail if the context does
not exactly match.

To resolve these issues, this commit pulls in as many trivial changes
from 4.5 (which we assume is nearly-identical to upstream) as possible.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Acked-by: Ping Cheng <ping.cheng@wacom.com>