Synchronize trivial changes into to 2.6.38
A number of minor changes have either been unintentionally made to the
2.6.38 tree or occurred as a result of not backporting all upstream changes.
These differences do not affect driver operation, but the clutter they
produce in the output of `diff` does serve to hide a number of "real"
semantic differences. They also make it more difficult to backport
upstream patches, as the patch command will fail if the context does
not exactly match.

To resolve these issues, this commit pulls in as many trivial changes
from 3.7 (which itself was synchronized in the prior patch) as possible.

Notable changes include:

  * Linebreak in each 'struct wacom_features' definition moved to occur
    after the 'distance_max' field like upstream does

  * The 'wacom_ids' table is reordered by PID like upstream

  * Maximum tablet coordinates of DTH2242 corrected

  * Several cases inside 'wacom_setup_input_capabilities' are moved to
    match their position within 3.7. Where fallthroughs are affected,
    identical behavior has been confirmed.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Acked-by: Ping Cheng <ping.cheng@wacom.com>