Add a special case for RHEL 7.4+
RHEL 7.4 ships both the USB module and the HID version.
We need to use the HID version if we do not want users
to manually quirk their devices.

Compile both modules and make sure we install them
properly.

Note that on these distributions, wacom_w8001.ko is compiled
twice, but given the 2 files in 3.7 and 3.17 are identical,
it doesn't hurt much.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>