Correct tilt range for MobileStudio/Intuos/Cintiq Pro devices
The pre-3.17 kernels set up their tilt axis ranges to be 0..127 rather
than -64..+63 like with the HID kernels. We need to add a correction for
this to the wacom_mspro_pen_irq function so that userspace properly
interprets the data. Without this fix, any physical tilt angle will be
interprted as being to the upper-left (and any amount of actual up/left
tilt angle will be saturated at -64 degrees).

Fixes: cd79aac597 ("backport: MobileStudio Pro to 3.7, 2.6.38, 2.6.36, 2.6.30")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Acked-by: Ping Cheng <ping.cheng@wacom.com>