Remove 2.6.36 kernel tree
The 2.6.36 tree is an odd duck, supporting an early version of the
multitouch protocol. This version of the protocol is different from
the final form used by 2.6.38+ and the linuxwacom-specific two-finger
faux-multitouch used by some devices in the 2.6.30 tree. Additionally,
before the integration of Travis-CI, this tree was often forgot about
when backporting, making it even further unlike its siblings. All
this makes continued maintainence of the tree particulary difficult.

Distributions which once used the 2.6.36 and 2.6.37 kernels are long
past their end of life (RIP openSUSE 11.4, 2014-09-27) so we can safely
remove this tree and never think about it again.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>