Implement support for multitouch value conflict resolution on 2.6.35+
Devices using the 2.6.30 wacom-specific multitouch protocol have to be
careful to ensure that two contacts don't use the same axis value. This
is because the second contact's value will be swallowed by the kernel.
The 2.6.36 kernel replaced direct-access to the value with an accessor
function (named "input_abs_get_val"), but our 2.6.30-compatible branch
kept using direct access, though only for pre-2.6.35 kernels (off-by-one
error on the version number).

This commit introduces an "input_abs_get_val" function for pre-2.6.36
kernels. This allows us to use the native API where available and
fallback to direct access when not. It also means that the fixups are
now applied to 2.6.35+ kernels.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>