2.6.30: Add missing oVid/oPid data to pen side of split devices
When oVid/oPid data was added for several "split" devices in afb48f975b,
the data was only added to the touch half. This means that the current
logic to link the two halves together will only work when the pen half
is detected first. This would cause arbitration issues if the probe
order is reversed and touch is detected first. By adding the oVid/oPid
data to the pen side as well, the logic will link the devices regardless
of probe order.

Fixes: afb48f975b ("Add support for WACOM_24HDT and WACOM_27QHDT touch sensors to 2.6.30 / 2.6.36")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>