2.6.30: Do not report acceleromter data from Cintiq 27QHD
The 2.6.38 and 3.7 kernels do not report accelerometer data from the
27QHD, likely because of the issues associated with reporting ABS_{X,Y,Z}
from a pad device that isn't seperate from the pen device. The X driver
should handle this (since it will place it into the pad's dedicated
channel) but tracking prox makes things tricky: either you always keep
the pad in prox like the 2.6.30 kernel was doing (which will possibly
be a source of actual issues) or you just don't bother reporting this
data like the 2.6.38 and 3.7 kernels.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>