3.17/Makefile.in: don't fail uninstall if a .ko file isn't present
The 3.17 path will now only build a driver named "hid-wacom.ko" if
the driver is installed on RHEL 7.4 or RHEL 7.5. Thus, we shouldn't
fail "make uninstall" if the "wacom.ko" that we normally build into
the "extra" directory is missing.

Also make sure we remove "hid-wacom.ko" on uninstall, and also don't
fail if it is missing as most other distributions will not build it.

Signed-off-by: Aaron Armstrong Skomra<skomra@gmail.com>