Add support for kernel module signing and enforcement
It is becoming more common for systems to only allow signed kernel modules
to be loaded (e.g. because the kernel is in "lockdown mode" due to secure
boot policies). Users who follow the standard configure / make / install
process find that the driver does not work after rebooting and often don't
know what is wrong.

This commit teaches the configure script and Makefiles a few new tricks:

  * Detection of when module signing is required by the system
  * Detection of existing key and cert used for module signing (Ubuntu
    only)
  * Ability to manually specify key, cert, and hash algorithm to be
    used for module signing
  * Abort configure if signing is required but not possible
  * New makefile target to add the signature prior to installing

Module signing is opportunistic by default. If a key and cert can be
found (or are provided to the configure script), the modules will be
signed -- even if not otherwise required. If a key and cert aren't
available then we will try to build an unsigned driver instead. If
the system is set up to require signed modules, we will cowardly
error out of the configure script unless the user explicitly requests
the driver to be left unsigned.

A phony "signature" Makefile target is provided which is called by
`make install` and which is available for developers who want/need
to sign the module without actually installing it. Users can use
`make && sudo make install` while developers can use `make && sudo
make signature`. If module signing is disabled the "signature"
target does nothing.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Acked-by: Ping Cheng <Ping.cheng@wacom.com>